FEAT: add crawl info by jav star code
