+++ .github/workflows/release.yml
@@ -3,7 +3,7 @@ name: Release
 on:
   push:
     tags:
-      - 'v*'
+      - "v*"
 
 jobs:
   build:
@@ -16,14 +16,13 @@ jobs:
       - name: Set up Go 1.x
         uses: actions/setup-go@v4
         with:
-          go-version: ^1.20
+          go-version: ^1.21
         id: go
 
       - run: go version
 
       - name: Get dependencies
-        run:
-          go get -v -t -d ./...
+        run: go get -v -t -d ./...
 
       - uses: goreleaser/goreleaser-action@v4
         with:
